Merge upstream release v0.1.8 (#6)
* Get updated cluster before removing finalizers (#1173)

If the cluster actuator's Delete() call updates the cluster in the
apiserver, there is a possibility that in-memory cluster object in the
cluster controller is now out of date. We either need to get-then-update
the cluster to remove the finalizer, or we need to use patch. The
version of controller-runtime we're using in release-0.1 does not
support patch in its generic client, so this change adds the
get-then-update logic, and wraps it inside a retry on conflict loop.

Signed-off-by: Andy Goldstein <goldsteina@vmware.com>

* Update lodash version to address CVE (#1156)

Signed-off-by: Andy Goldstein <goldsteina@vmware.com>

* Update Netlify config for v1alpha1 subdomain (#1176)

* Release v0.1.8 (#1183)